requirements.txt
google-generativeai>=0.3.0
requests>=2.25.0
python-dotenv>=0.19.0
pydantic>=1.8.0
pytest>=6.0.0
pytest-asyncio>=0.18.0
aiohttp>=3.8.0
nest-asyncio>=1.5.0
Jinja2>=3.0.0
